Revert "cleanliness: cast once at the switch, not at all 74 case labels"
This reverts commit 58b23518a55f8e58b17e5cf3259b155e2b4d4545.

+++ src/DDrawMgr/DirectDrawMgr.cpp
@@ -70,60 +70,60 @@ void CDDrawPtrCollections::GetErrorString(char* file, i32 line, i32 hr) {
     sprintf(szCode, "Unknown Error Code");
     strcpy(szLine, g_emptyString);
 
-    switch (static_cast<u32>(hr)) {
-        case DDERR_UNSUPPORTED:
+    switch (hr) {
+        case static_cast<i32>(DDERR_UNSUPPORTED):
             strcpy(szCode, "DDERR_UNSUPPORTED");
             strcpy(szMsg, "Action not supported");
             break;
-        case DDERR_GENERIC:
+        case static_cast<i32>(DDERR_GENERIC):
             strcpy(szCode, "DDERR_GENERIC");
             strcpy(szMsg, "Generic failure");
             break;
-        case DDERR_OUTOFMEMORY:
+        case static_cast<i32>(DDERR_OUTOFMEMORY):
             strcpy(szCode, "DDERR_OUTOFMEMORY");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_INVALIDPARAMS:
+        case static_cast<i32>(DDERR_INVALIDPARAMS):
             strcpy(szCode, "DDERR_INVALIDPARAMS");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_INVALIDCAPS:
+        case static_cast<i32>(DDERR_INVALIDCAPS):
             strcpy(szCode, "DDERR_INVALIDCAPS");
             strcpy(szMsg, "One or more of the caps bits passed to the callback are incorrect");
             break;
-        case DDERR_INVALIDMODE:
+        case static_cast<i32>(DDERR_INVALIDMODE):
             strcpy(szCode, "DDERR_INVALIDMODE");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_INVALIDOBJECT:
+        case static_cast<i32>(DDERR_INVALIDOBJECT):
             strcpy(szCode, "DDERR_INVALIDOBJECT");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_INVALIDPIXELFORMAT:
+        case static_cast<i32>(DDERR_INVALIDPIXELFORMAT):
             strcpy(szCode, "DDERR_INVALIDPIXELFORMAT");
             strcpy(szMsg, "Pixel format was invalid as specified.");
             break;
-        case DDERR_INVALIDRECT:
+        case static_cast<i32>(DDERR_INVALIDRECT):
             strcpy(szCode, "DDERR_INVALIDRECT");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_LOCKEDSURFACES:
+        case static_cast<i32>(DDERR_LOCKEDSURFACES):
             strcpy(szCode, "DDERR_LOCKEDSURFACES");
             strcpy(szMsg, "No message");
             break;
-        case 0x887600aa:
+        case static_cast<i32>(0x887600aa):
             strcpy(szCode, "DDERR_NO3D");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOALPHAHW:
+        case static_cast<i32>(DDERR_NOALPHAHW):
             strcpy(szCode, "DDERR_NOALPHAHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOCOLORCONVHW:
+        case static_cast<i32>(DDERR_NOCOLORCONVHW):
             strcpy(szCode, "DDERR_NOCOLORCONVHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOCOOPERATIVELEVELSET:
+        case static_cast<i32>(DDERR_NOCOOPERATIVELEVELSET):
             strcpy(szCode, "DDERR_NOCOOPERATIVELEVELSET");
             strcpy(
                 szMsg,
@@ -131,15 +131,15 @@ void CDDrawPtrCollections::GetErrorString(char* file, i32 line, i32 hr) {
                 "called"
             );
             break;
-        case DDERR_NOEXCLUSIVEMODE:
+        case static_cast<i32>(DDERR_NOEXCLUSIVEMODE):
             strcpy(szCode, "DDERR_NOEXCLUSIVEMODE");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOGDI:
+        case static_cast<i32>(DDERR_NOGDI):
             strcpy(szCode, "DDERR_NOGDI");
             strcpy(szMsg, "There is no GDI present");
             break;
-        case DDERR_NOMIRRORHW:
+        case static_cast<i32>(DDERR_NOMIRRORHW):
             strcpy(szCode, "DDERR_NOMIRRORHW");
             strcpy(
                 szMsg,
@@ -147,85 +147,85 @@ void CDDrawPtrCollections::GetErrorString(char* file, i32 line, i32 hr) {
                 "available."
             );
             break;
-        case DDERR_NOTFOUND:
+        case static_cast<i32>(DDERR_NOTFOUND):
             strcpy(szCode, "DDERR_NOTFOUND");
             strcpy(szMsg, "Request item was not found");
             break;
-        case DDERR_NOOVERLAYHW:
+        case static_cast<i32>(DDERR_NOOVERLAYHW):
             strcpy(szCode, "DDERR_NOOVERLAYHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NORASTEROPHW:
+        case static_cast<i32>(DDERR_NORASTEROPHW):
             strcpy(szCode, "DDERR_NORASTEROPHW");
             strcpy(szMsg, "No message");
             break;
         // retail prints "DDERR_NOROTATEHW" here; the value is the SDK's DDERR_NOROTATIONHW.
-        case DDERR_NOROTATIONHW:
+        case static_cast<i32>(DDERR_NOROTATIONHW):
             strcpy(szCode, "DDERR_NOROTATEHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOSTRETCHHW:
+        case static_cast<i32>(DDERR_NOSTRETCHHW):
             strcpy(szCode, "DDERR_NOSTRETCHHW");
             strcpy(szMsg, "No message");
             break;
-        case 0x88760140:
+        case static_cast<i32>(0x88760140):
             strcpy(szCode, "DDERR_NOT8BITCOLOR");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOTEXTUREHW:
+        case static_cast<i32>(DDERR_NOTEXTUREHW):
             strcpy(szCode, "DDERR_NOTEXTUREHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOVSYNCHW:
+        case static_cast<i32>(DDERR_NOVSYNCHW):
             strcpy(szCode, "DDERR_NOVSYNCHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOZBUFFERHW:
+        case static_cast<i32>(DDERR_NOZBUFFERHW):
             strcpy(szCode, "DDERR_NOZBUFFERHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_OUTOFCAPS:
+        case static_cast<i32>(DDERR_OUTOFCAPS):
             strcpy(szCode, "DDERR_OUTOFCAPS");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_OUTOFVIDEOMEMORY:
+        case static_cast<i32>(DDERR_OUTOFVIDEOMEMORY):
             strcpy(szCode, "DDERR_OUTOFVIDEOMEMORY");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_PALETTEBUSY:
+        case static_cast<i32>(DDERR_PALETTEBUSY):
             strcpy(szCode, "DDERR_PALETTEBUSY");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_SURFACEBUSY:
+        case static_cast<i32>(DDERR_SURFACEBUSY):
             strcpy(szCode, "DDERR_SURFACEBUSY");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_SURFACEISOBSCURED:
+        case static_cast<i32>(DDERR_SURFACEISOBSCURED):
             strcpy(szCode, "DDERR_SURFACEISOBSCURED");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_SURFACELOST:
+        case static_cast<i32>(DDERR_SURFACELOST):
             strcpy(szCode, "DDERR_SURFACELOST");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_SURFACENOTATTACHED:
+        case static_cast<i32>(DDERR_SURFACENOTATTACHED):
             strcpy(szCode, "DDERR_SURFACENOTATTACHED");
             strcpy(szMsg, "The requested surface is not attached");
             break;
-        case DDERR_TOOBIGSIZE:
+        case static_cast<i32>(DDERR_TOOBIGSIZE):
             strcpy(szCode, "DDERR_TOOBIGSIZE");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_TOOBIGWIDTH:
+        case static_cast<i32>(DDERR_TOOBIGWIDTH):
             strcpy(szCode, "DDERR_TOOBIGWIDTH");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_VERTICALBLANKINPROGRESS:
+        case static_cast<i32>(DDERR_VERTICALBLANKINPROGRESS):
             strcpy(szCode, "DDERR_VERTICALBLANKINPROGRESS");
             strcpy(szMsg, "No message");
             break;
         // retail prints "DDERR_WASTILLDRAWING" here; the value is the SDK's DDERR_WASSTILLDRAWING.
-        case DDERR_WASSTILLDRAWING:
+        case static_cast<i32>(DDERR_WASSTILLDRAWING):
             strcpy(szCode, "DDERR_WASTILLDRAWING");
             strcpy(
                 szMsg,
@@ -233,40 +233,40 @@ void CDDrawPtrCollections::GetErrorString(char* file, i32 line, i32 hr) {
                 "incomplete"
             );
             break;
-        case DDERR_NODIRECTDRAWHW:
+        case static_cast<i32>(DDERR_NODIRECTDRAWHW):
             strcpy(szCode, "DDERR_NODIRECTDRAWHW");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_DIRECTDRAWALREADYCREATED:
+        case static_cast<i32>(DDERR_DIRECTDRAWALREADYCREATED):
             strcpy(szCode, "DDERR_DIRECTDRAWALREADYCREATED");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_XALIGN:
+        case static_cast<i32>(DDERR_XALIGN):
             strcpy(szCode, "DDERR_XALIGN");
             strcpy(szMsg, "Rectangle provided was not horizontally aligned on a DWORD boundary");
             break;
-        case DDERR_HWNDSUBCLASSED:
+        case static_cast<i32>(DDERR_HWNDSUBCLASSED):
             strcpy(szCode, "DDERR_HWNDSUBCLASSED");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_HWNDALREADYSET:
+        case static_cast<i32>(DDERR_HWNDALREADYSET):
             strcpy(szCode, "DDERR_HWNDALREADYSET");
             strcpy(szMsg, "No message");
             break;
-        case DDERR_NOPALETTEHW:
+        case static_cast<i32>(DDERR_NOPALETTEHW):
             strcpy(szCode, "DDERR_NOPALETTEHW");
             strcpy(szMsg, "No hardware support for 16 or 256 color palettes");
             break;
-        case DDERR_PRIMARYSURFACEALREADYEXISTS:
+        case static_cast<i32>(DDERR_PRIMARYSURFACEALREADYEXISTS):
             strcpy(szCode, "DDERR_PRIMARYSURFACEALREADYEXISTS");
             strcpy(szMsg, "This process already has created a primary surface");
             break;
-        case DDERR_EXCLUSIVEMODEALREADYSET:
+        case static_cast<i32>(DDERR_EXCLUSIVEMODEALREADYSET):
             strcpy(szCode, "DDERR_EXCLUSIVEMODEALREADYSET");
             strcpy(szMsg, "No message");
             break;
         // retail prints "DDERR_LOCKEDSURFACES" here; the value is the SDK's DDERR_NOTLOCKED.
-        case DDERR_NOTLOCKED:
+        case static_cast<i32>(DDERR_NOTLOCKED):
             strcpy(szCode, "DDERR_LOCKEDSURFACES");
             strcpy(szMsg, "No message");
             break;
